cloud-function-embeddings/retrieve-embeddings/requirements.txt
functions-framework==3.*
google-cloud-storage==2.10.0
flask==3.0.0
